experiments/requirements.txt
langchain[google-genai]==0.3.27
langchain-community==0.3.27
faiss-cpu==1.11.0.post1
pydantic==2.11.7
transformers==4.45.2